ad54a9b Change new constants in txdb.h to int64_t (Wladimir J. van der Laan)

+++ src/txdb.h
@@ -19,11 +19,11 @@ class CCoins;
 class uint256;
 
 // -dbcache default (MiB)
-static const int nDefaultDbCache = 100;
+static const int64_t nDefaultDbCache = 100;
 // max. -dbcache in (MiB)
-static const int nMaxDbCache = sizeof(void*) > 4 ? 4096 : 1024;
+static const int64_t nMaxDbCache = sizeof(void*) > 4 ? 4096 : 1024;
 // min. -dbcache in (MiB)
-static const int nMinDbCache = 4;
+static const int64_t nMinDbCache = 4;
 
 /** CCoinsView backed by the LevelDB coin database (chainstate/) */
 class CCoinsViewDB : public CCoinsView
